requirements.txt
google-genai
pandas
python-dotenv
tenacity
streamlit
plotly